+++ CHANGELOG.md
@@ -26,6 +26,10 @@ Given a version number `MAJOR.MINOR.PATCH`, we increment the:
 
 - Add `terramate.stack.parent.id` metadata to stacks that are part of a parent-child hierarchy.
 
+### Fixed
+
+- Nested `map` blocks not rendered if the `value` block contain no attributes.
+
 ## v0.11.1
 
 ### Added